question/requirements.txt
streamlit>=1.28.0
pypdf>=3.17.0
openai>=1.12.0
python-dotenv>=1.0.0
reportlab>=4.0.0
